Create ResultPanelSubscriptionProvider
