+++ client/src/components/ResultPanel/ResultPanel.ts
